Clean up some p4est-related code.
Specifically, commit daf3146 (via #3625) introduced a dimension dependent dispatch
mechanism that allowed writing some code in a more generic way. The idea was right,
but the approach duplicated the dispatching because we already have a dispatching
mechanism via the internal::p4est::functions classes.

This patch simply merges the two approaches. It also allows to move some
code out of distributed/tria.cc into distributed/p4est_wrappers.cc, at the
cost of some code churn. The patch does not actually change any kind of
functionality -- it just moves things.
