+++ source/distributed/p4est_wrappers.inst.in
@@ -58,3 +58,28 @@ for (deal_II_dimension : DIMENSIONS)
     \}
 #   endif // DEAL_II_WITH_P4EST
 }
+
+
+for (deal_II_dimension, deal_II_space_dimension : DIMENSIONS)
+{
+#   ifdef DEAL_II_WITH_P4EST
+
+    namespace internal
+    \{
+    namespace p4est
+    \{
+#if     deal_II_dimension > 1
+#if       deal_II_space_dimension >= deal_II_dimension
+    template
+    void
+    functions<deal_II_dimension>::
+    iterate<deal_II_space_dimension>(dealii::internal::p4est::types<deal_II_dimension>::forest *parallel_forest,
+                                     dealii::internal::p4est::types<deal_II_dimension>::ghost *parallel_ghost,
+                                     void *user_data);
+#endif
+#endif
+    \}
+    \}
+#   endif // DEAL_II_WITH_P4EST
+}
+
